Fully finished all algorithms and formatted code
Fully completed A* and DFS, put code into PEP8 standards (pycodestyle) and formatted the code to achieve an adequate pylint score (9.6+ for each).

Additionally, added more comments and made all docstrings compliant with PEP8.

Placed the recursive DFS into a legacy folder as it isn't going to be used, but don't want to get rid of it.